requirements.txt
google-generativeai
pydantic
python-dotenv
autogen
autogen-core
autogen-agentchat
ag2[gemini]